datastructure: hashtable: increase capacity return type
if increase capacity can't allocate stuff due to memory pressure then,
this is "okay" (not fatal). The reason for that is because the worst
case scenario would be to have additional conflicts on the hashtable
which would tend to O(n). However, this is way better than failing the
whole operation from my humble point of view.