parser: semantic: pass a reference to databse to semantic_analyse fnc
It became clear to me that in order to fully perform semantic analysis I
would have to have a reference to the current database so I can fetch
tables from the database dictionary to check for some edge cases.

This implies that we would have to lock tables before and during
semantic analysis - but that's a problem to be solved when transaction
is implemented. I got tired of anticipating the transactions
implementation... I will cross that bridge once I come to it.