parser: semantic: div by 0 won't be handled in the semantic phase
to my great surprise, 1/0 is NULL in most database engines I played
with. That makes this not a semantic problem to fix but an edge case to
be handled in other phases.

So I added TODOs on the executor test to make sure I don't forget it but
most likely this will have to be handled in the optimisation phase so
execution phase is all about moving data around