parser: remove support for subqueries and SELECT INTO for SELECT stmts
SELECT INTO doesn't make much sense for MidoriDB as I have no intention
of supporting stored procedures or routines that can dump data to a file
like MySQL has.

Subqueries was a tough choice. I wanted to implement it but given that
this is a toy db, I don't think it's worth the effort for a pet project.
Being brutally honest, when I implement the SELECT stmts I would have
achieved the peak knowledge for that operation, so implementing subquery
wouldn't make me learn anything new really. Anyway, if anyone wants to
implement it and send a PR, by all means, have fun :)